[Refactor] 게임페이지 리팩토링 #447
