--- components/game/GameResultItem.tsx
@@ -1,48 +1,0 @@
-import { useSetRecoilState } from 'recoil';
-import { Game } from 'types/gameTypes';
-import { clickedGameItemState } from 'utils/recoil/game';
-import GameResultBigScore from 'components/game/big/GameResultBigScore';
-import GameResultSmallScore from 'components/game/small/GameResultSmallScore';
-import GameResultBigTeam from 'components/game/big/GameResultBigTeam';
-import GameResultSmallTeam from 'components/game/small/GameResultSmallTeam';
-import styles from 'styles/game/GameResultItem.module.scss';
-interface GameResultItemProps {
-  game: Game;
-  isBig: boolean;
-}
-
-export default function GameResultItem({ game, isBig }: GameResultItemProps) {
-  const { mode, team1, team2, status, time, gameId } = game;
-  const setClickedItemId = useSetRecoilState(clickedGameItemState);
-  return (
-    <div
-      className={`${isBig ? styles.bigContainer : styles.smallContainer}
-			${mode === 'normal' ? styles.normal : styles.rank}`}
-      onClick={() => setClickedItemId(gameId)}
-    >
-      {isBig ? (
-        <>
-          <GameResultBigTeam team={team1} />
-          <GameResultBigScore
-            mode={mode}
-            status={status}
-            time={time}
-            scoreTeam1={team1.score}
-            scoreTeam2={team2.score}
-          />
-          <GameResultBigTeam team={team2} />
-        </>
-      ) : (
-        <>
-          <GameResultSmallTeam team={team1} userLeft={true} />
-          <GameResultSmallScore
-            mode={mode}
-            scoreTeam1={team1.score}
-            scoreTeam2={team2.score}
-          />
-          <GameResultSmallTeam team={team2} userLeft={false} />
-        </>
-      )}
-    </div>
-  );
-}
