+++ components/game/big/GameResultBigItem.tsx
@@ -1,0 +1,32 @@
+import { useSetRecoilState } from 'recoil';
+import { Game } from 'types/gameTypes';
+import { clickedGameItemState } from 'utils/recoil/game';
+import GameResultBigScore from 'components/game/big/GameResultBigScore';
+import GameResultBigTeam from 'components/game/big/GameResultBigTeam';
+import styles from 'styles/game/GameResultItem.module.scss';
+
+interface GameResultBigItemProps {
+  game: Game;
+}
+
+export default function GameResultBigItem({ game }: GameResultBigItemProps) {
+  const { mode, team1, team2, status, time, gameId } = game;
+  const setClickedItemId = useSetRecoilState(clickedGameItemState);
+  return (
+    <div
+      className={`${styles.bigContainer}
+			${mode === 'normal' ? styles.normal : styles.rank}`}
+      onClick={() => setClickedItemId(gameId)}
+    >
+      <GameResultBigTeam team={team1} />
+      <GameResultBigScore
+        mode={mode}
+        status={status}
+        time={time}
+        scoreTeam1={team1.score}
+        scoreTeam2={team2.score}
+      />
+      <GameResultBigTeam team={team2} />
+    </div>
+  );
+}
